Split results into their own file
