+++ src/larceler/LarStandaloneRunner.hh
@@ -11,9 +11,7 @@
 
 #include "corecel/Macros.hh"
 #include "corecel/cont/Span.hh"
-#include "corecel/math/Quantity.hh"
 #include "geocel/Types.hh"
-#include "celeritas/UnitTypes.hh"
 #include "celeritas/optical/DetectorData.hh"
 
 namespace sim
