+++ test/larceler/CMakeLists.txt
@@ -17,6 +17,7 @@ celeritas_setup_tests(SERIAL
 
 celeritas_add_test(LarStandaloneRunner.test.cc
   LINK_LIBRARIES lardataobj::Simulation
+  SOURCES RunnerResults.cc
 )
 celeritas_add_test(LarDataReader.test.cc LINK_LIBRARIES Celeritas::celeritas)
